+++ src/main/java/de/maxhenkel/car/integration/waila/HUDHandlerCars.java
@@ -1,7 +1,7 @@
 package de.maxhenkel.car.integration.waila;
-/*
+
 import de.maxhenkel.car.entity.car.base.EntityGenericCar;
-import de.maxhenkel.tools.MathTools;
+import de.maxhenkel.corelib.math.MathUtils;
 import mcp.mobius.waila.Waila;
 import mcp.mobius.waila.api.IEntityAccessor;
 import mcp.mobius.waila.api.IEntityComponentProvider;
@@ -9,7 +9,10 @@ import mcp.mobius.waila.api.IPluginConfig;
 import mcp.mobius.waila.api.ITaggableList;
 import mcp.mobius.waila.utils.ModIdentification;
 import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.*;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraftforge.fluids.FluidStack;
 
 import java.util.List;
@@ -19,11 +22,11 @@ public class HUDHandlerCars implements IEntityComponentProvider {
     static final HUDHandlerCars INSTANCE = new HUDHandlerCars();
 
     @Override
-    public void appendHead(List<ITextComponent> tip, IEntityAccessor accessor, IPluginConfig config) {
-        ITaggableList<ResourceLocation, ITextComponent> tooltip = (ITaggableList<ResourceLocation, ITextComponent>) tip;
-        tooltip.setTag(PluginCar.OBJECT_NAME_TAG, new StringTextComponent(String.format(Waila.CONFIG.get().getFormatting().getEntityName(), accessor.getEntity().getDisplayName().getFormattedText())));
+    public void appendHead(List<ITextComponent> t, IEntityAccessor accessor, IPluginConfig config) {
+        ITaggableList<ResourceLocation, ITextComponent> tooltip = (ITaggableList<ResourceLocation, ITextComponent>) t;
+        tooltip.setTag(PluginCar.OBJECT_NAME_TAG, new StringTextComponent(String.format(Waila.CONFIG.get().getFormatting().getEntityName(), accessor.getEntity().getDisplayName().getString())));
         if (config.get(PluginCar.CONFIG_SHOW_REGISTRY)) {
-            tooltip.setTag(PluginCar.REGISTRY_NAME_TAG, new StringTextComponent(String.format(Waila.CONFIG.get().getFormatting().getRegistryName(), accessor.getEntity().getType().getRegistryName().toString())));
+            tooltip.setTag(PluginCar.REGISTRY_NAME_TAG, new StringTextComponent(accessor.getEntity().getType().getRegistryName().toString()).func_240699_a_(TextFormatting.GRAY));
         }
     }
 
@@ -44,7 +47,7 @@ public class HUDHandlerCars implements IEntityComponentProvider {
 
         float damage = car.getDamage();
         if (damage > 0F) {
-            ITextComponent dmg = new TranslationTextComponent("tooltip.waila.car.damage", MathTools.round(damage, 2));
+            ITextComponent dmg = new TranslationTextComponent("tooltip.waila.car.damage", MathUtils.round(damage, 2));
             tooltip.add(dmg);
         }
     }
@@ -53,4 +56,5 @@ public class HUDHandlerCars implements IEntityComponentProvider {
     public void appendTail(List<ITextComponent> tooltip, IEntityAccessor accessor, IPluginConfig config) {
         tooltip.add(new StringTextComponent(String.format(Waila.CONFIG.get().getFormatting().getModName(), ModIdentification.getModInfo(accessor.getEntity()).getName())));
     }
-}*/
+
+}